worker_threads: prime node:console before the lazy getter; move test to own file
require('node:console') is export default globalThis.console; evaluating it
after the lazy console accessor is installed would cache a plain Console
instance that lacks .Console/.write/[Symbol.asyncIterator]. Capture via
require first so the module registry caches the native object.

Move the lazy-accessor regression test into its own file so the gate can run
it cleanly under ASAN (worker_threads.test.ts has subprocess-spawning cases
that exceed the 5s default under ASAN).
